serial commit number in version string; wipe endpoint in test makefile; added missing test number; added SERVER directive to test makefile allowing for bigger kmers; separated read and write phases in run_test for the sake of memory consumption

+++ test/run_test.py
@@ -11,8 +11,9 @@ print dir(kmer_set)
 kmer_set.version()
 
 def main():
-	kmer_lens    = [ int(k) for k in sys.argv[1].split() ]
-	test_indexes = [ int(i) for i in sys.argv[2].split() ]
+	exec_type    =                   sys.argv[1]
+	kmer_lens    = [ int(k) for k in sys.argv[2].split() ]
+	test_indexes = [ int(i) for i in sys.argv[3].split() ]
 
 	outs = []
 	for kmer_len in kmer_lens:
@@ -25,22 +26,24 @@ def main():
 			print fa, km
 			outs.append(km)
 
-			if os.path.exists(km):
-				os.remove(km)
+			if exec_type == 'write':
+				if os.path.exists(km):
+					os.remove(km)
+	
+				ek = kmer_set.extract_kmers(kmer_len)
+				ek.read_one_liner(fa)
+				ek.save_kmer_db(km)
+				print 'FINISHED SAVING'
 
-			ek = kmer_set.extract_kmers(kmer_len)
-			ek.read_one_liner(fa)
-			ek.save_kmer_db(km)
-			print 'FINISHED SAVING'
-
-			if os.path.exists(km):
-				ek2 = kmer_set.extract_kmers(kmer_len)
-				ek2.read_kmer_db(km)
-				re = ek.get_kmer_db()
-
-				print 'RE:', (re if (kmer_len <= 9) else '')
-
-				print 'FINISHED READING'
+			elif exec_type == 'read':
+				if os.path.exists(km):
+					ek2 = kmer_set.extract_kmers(kmer_len)
+					ek2.read_kmer_db(km)
+					re = ek2.get_kmer_db()
+	
+					print 'RE:', (re if (kmer_len <= 9) else '')
+	
+					print 'FINISHED READING'
 
 if __name__ == "__main__":
 	main()
